refactor: `core/cert` --> `core/downloader`
