+++ core/builder/cipher_builder.go
@@ -4,8 +4,8 @@ import (
 	"crypto/x509"
 
 	"github.com/wechatpay-apiv3/wechatpay-go/core"
-	"github.com/wechatpay-apiv3/wechatpay-go/core/cert"
 	"github.com/wechatpay-apiv3/wechatpay-go/core/cipher/encryptors"
+	"github.com/wechatpay-apiv3/wechatpay-go/core/downloader"
 )
 
 func BuildWechatPayEncryptorWithCertificateList(l []*x509.Certificate) *encryptors.WechatPayEncryptor {
@@ -13,7 +13,7 @@ func BuildWechatPayEncryptorWithCertificateList(l []*x509.Certificate) *encrypto
 }
 
 func BuildWechatPayEncryptorWithDownloaderMgr(
-	mgr *cert.CertificateDownloaderMgr, mchID string,
+	mgr *downloader.CertificateDownloaderMgr, mchID string,
 ) *encryptors.WechatPayEncryptor {
 	return encryptors.NewWechatPayEncryptor(mgr.GetCertificateVisitor(mchID))
 }
