+++ core/downloader/downloader.go
@@ -1,4 +1,4 @@
-package cert
+package downloader
 
 import (
 	"context"
