+++ core/downloader/downloader_mgr_singleton.go
@@ -1,4 +1,4 @@
-package cert
+package downloader
 
 import (
 	"context"
